refactored to use settings file
now uses settings.txt which is generated in local directory to hand seed which is also provided in the window title. also refactored to use Lib.hpp for requesting loaded seed and handling settings. changed  Makefile to use c++20 as need c++17 at least for the lazymode checking if a file exists